+++ DC-EMS Android Project/.gitignore
@@ -7,6 +7,7 @@
 /.idea/workspace.xml
 /.idea/navEditor.xml
 /.idea/assetWizardSettings.xml
+/.idea/discord.xml
 .DS_Store
 /build
 /captures
