+++ DC-EMS Android Project/app/src/androidTest/java/com/yangPentingLulus/DCEMS/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.DCEMS.hello_world
+package com.yangPentingLulus.DCEMS
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
