+++ DC-EMS Android Project/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.DCEMS.hello_world">
+    package="com.yangPentingLulus.DCEMS">
 
     <application
 
@@ -10,8 +10,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity"/>
-        <activity android:name=".SplashScreenActivity">
+        <activity android:name="com.yangPentingLulus.DCEMS.MainActivity"/>
+        <activity android:name="com.yangPentingLulus.DCEMS.SplashScreenActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
