+++ DC-EMS Android Project/app/src/main/java/com/yangPentingLulus/DCEMS/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.DCEMS.hello_world
+package com.yangPentingLulus.DCEMS
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
