+++ DC-EMS Android Project/app/src/main/java/com/yangPentingLulus/DCEMS/SplashScreenActivity.kt
@@ -1,4 +1,4 @@
-package com.DCEMS.hello_world
+package com.yangPentingLulus.DCEMS
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
