+++ DC-EMS Android Project/app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">Helloworld</string>
+    <string name="app_name">DCEMS</string>
     <string name="image_splashscreen">image_splashscreen</string>
 </resources>
