+++ DC-EMS Android Project/app/src/test/java/com/yangPentingLulus/DCEMS/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.DCEMS.hello_world
+package com.yangPentingLulus.DCEMS
 
 import org.junit.Test
